+++ variants/cardputer_adv/companion_radio/ui-cardputer/CardputerUITask.cpp
@@ -143,6 +143,22 @@ void CardputerUITask::shutdown(bool restart) {
   }
 }
 
+// In case if input has utf-8 multibyte characters
+void CardputerUITask::removeLastStringChar(String &str) {
+  unsigned int len = str.length();
+  if (len == 0) {
+    return;
+  }
+
+  unsigned int bytesToRemove = 1;
+
+  while (len - bytesToRemove > 0 && ((uint8_t)str.charAt(len - bytesToRemove) & 0xC0) == 0x80) {
+    bytesToRemove++;
+  }
+
+  str.remove(len - bytesToRemove, bytesToRemove);
+}
+
 void CardputerUITask::loop() {
   _board->getSpeaker()->processQueue();
 
@@ -207,7 +223,6 @@ bool CardputerUITask::turnDisplayOn() {
       _display->turnOn();
       return true;
     }
-
   }
   return false;
 }