+++ variants/cardputer_adv/companion_radio/ui-cardputer/screen/MainScreen.h
@@ -49,8 +49,6 @@ class MainScreen : public CardputerScreen {
 
   int getChannelCount();
   void sendChatMessage();
-  void chatInputRemoveLastChar();
-
   void renderStatusIcons();
 
   int renderFirstPage();